+++ src/main/java/org/example/newsfeedproject/user/dto/SessionUserDto.java
@@ -1,4 +1,4 @@
-package org.example.newsfeedproject.dto;
+package org.example.newsfeedproject.user.dto;
 
 public class SessionUserDto {
